+++ Tracker/Tracker/Trackers/List/Views/TrackersViewController.swift
@@ -138,186 +138,9 @@ extension TrackersViewController: UICollectionViewDelegate {
     }
 }
 
-final class TrackerCell: UICollectionViewCell {
-    
-    static let buttonRadius: CGFloat = 34
-        
-    private let backgroundColoredView: UIView = {
-        let coloredView = UIView()
-        coloredView.backgroundColor = .red
-        coloredView.layer.cornerRadius = Constant.cornerRadius
-        return coloredView
-    }()
-    
-    private let plusButton: UIButton = {
-        let button = UIButton(type: .custom)
-        button.backgroundColor = .red
-        button.setImage(UIImage(systemName: "plus"), for: .normal)
-        button.tintColor = .white
-        button.layer.cornerRadius = TrackerCell.buttonRadius / 2
-        button.clipsToBounds = true
-        return button
-    }()
-    
-    private let countLabel: UILabel = {
-        let label = UILabel()
-        label.text = "1 день"
-        return label
-    }()
-    
-    private let nameTextView: UITextView = {
-        let textView = UITextView()
-        textView.font = UIFont.systemFont(ofSize: 12)
-        textView.text = "Поливать растения"
-        textView.isScrollEnabled = false
-        textView.isEditable = false
-        textView.isSelectable = false
-        textView.textContainer.maximumNumberOfLines = 2
-        textView.backgroundColor = .clear
-        textView.textColor = .white
-        
-        return textView
-    }()
-    
-    private let emogiView = EmojiView()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-
-        contentView.addSubviews([
-            backgroundColoredView,
-            plusButton,
-            countLabel,
-            nameTextView,
-            emogiView
-        ])
-        
-        NSLayoutConstraint.activate([
-            backgroundColoredView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor),
-            backgroundColoredView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor),
-            backgroundColoredView.topAnchor.constraint(equalTo: contentView.topAnchor),
-            backgroundColoredView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -58),
-            
-            plusButton.widthAnchor.constraint(equalToConstant: Self.buttonRadius),
-            plusButton.heightAnchor.constraint(equalToConstant: Self.buttonRadius),
-            plusButton.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -12),
-            plusButton.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -16),
-            
-            countLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 12),
-            countLabel.centerYAnchor.constraint(equalTo: plusButton.centerYAnchor),
-            
-            nameTextView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 44),
-            nameTextView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 12),
-            nameTextView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -12),
-            nameTextView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -12),
-        ])
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func update(tracker: Tracker) {
-        nameTextView.text = tracker.name
-        backgroundColoredView.backgroundColor = tracker.color.uiColor
-        plusButton.backgroundColor = tracker.color.uiColor
-        emogiView.setEmoji(tracker.icon)
-    }
-}
-
-final class HeaderView: UICollectionReusableView {
-    let titleLabel = UILabel()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        titleLabel.textColor = .black
-        titleLabel.font = UIFont.boldSystemFont(ofSize: 18)
-        
-        addSubview(titleLabel)
-        
-        titleLabel.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 12),
-            titleLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
-            titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 16),
-            titleLabel.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -12)
-        ])
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-final class TwoColumnFlowLayout: UICollectionViewFlowLayout {
-    override func prepare() {
-        super.prepare()
-        
-        guard let collectionView = collectionView else { return }
-        
-        let availableWidth = collectionView.bounds.width - sectionInset.left - sectionInset.right - minimumInteritemSpacing
-        let itemWidth = availableWidth / 2
-        
-        itemSize = CGSize(width: itemWidth, height: 148)
-        minimumLineSpacing = 0
-    }
-}
-
 extension TrackersViewController: UICollectionViewDelegateFlowLayout {
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
         return CGSize(width: collectionView.bounds.width, height: 46)
-    }
-}
-
-class EmojiView: UIView {
-    let roundedView: UIView = {
-        let roundedView = UIView()
-        roundedView.backgroundColor = UIColor(white: 1, alpha: 0.3)
-        return roundedView
-    }()
-    
-    private let emojiLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont.systemFont(ofSize: 12)
-        label.textAlignment = .center
-        return label
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupView()
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-        setupView()
-    }
-    
-    private func setupView() {
-        addSubviews([
-            roundedView,
-            emojiLabel
-        ])
-        
-        
-        NSLayoutConstraint.activate([
-            emojiLabel.centerXAnchor.constraint(equalTo: roundedView.centerXAnchor),
-            emojiLabel.centerYAnchor.constraint(equalTo: roundedView.centerYAnchor),
-            
-            roundedView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 12),
-            roundedView.topAnchor.constraint(equalTo: topAnchor, constant: 12),
-            roundedView.heightAnchor.constraint(equalToConstant: 24),
-            roundedView.widthAnchor.constraint(equalToConstant: 24)
-        ])
-        
-        roundedView.layer.cornerRadius = 12
-        roundedView.layer.masksToBounds = true
-
-    }
-    
-    func setEmoji(_ emoji: Character) {
-        emojiLabel.text = String(emoji)
     }
 }
